Add dataprocessing with worker processing (#12), cleanup repo structure (#11)
