+++ asim/script/builders/data_processor_builder.py
@@ -1,0 +1,27 @@
+import logging
+from typing import List
+
+from hydra.utils import instantiate
+from nuplan.planning.script.builders.utils.utils_type import validate_type
+from omegaconf import DictConfig
+
+from asim.dataset.dataset_specific.raw_data_processor import RawDataProcessor
+
+logger = logging.getLogger(__name__)
+
+
+def build_data_processor(cfg: DictConfig) -> List[RawDataProcessor]:
+    """
+    Builds the worker.
+    :param cfg: DictConfig. Configuration that is used to run the experiment.
+    :return: Instance of WorkerPool.
+    """
+    logger.info("Building RawDataProcessor...")
+    instantiated_datasets: List[RawDataProcessor] = []
+    for dataset_type in cfg.values():
+        processor: RawDataProcessor = instantiate(dataset_type)
+        validate_type(processor, RawDataProcessor)
+        instantiated_datasets.append(processor)
+
+    logger.info("Building RawDataProcessor...DONE!")
+    return instantiated_datasets
