+++ asim/script/config/dataset_caching/default_dataset_caching.yaml
@@ -4,6 +4,7 @@ hydra:
   output_subdir: ${output_dir}/code/hydra           # Store hydra's config breakdown here for debugging
   searchpath:                                       # Only <exp_dir> in these paths are discoverable
     - pkg://asim.script.config.common
+    - pkg://asim.script.config
   job:
     chdir: False
 
@@ -12,7 +13,8 @@ defaults:
   - default_experiment
   - default_dataset_paths
   - _self_
+  - datasets:
+    # - nuplan_mini_dataset
+    - carla_dataset
 
 force_feature_computation: True
-
-datasets: []
