+++ asim/script/config/datasets/carla_dataset.yaml
@@ -1,0 +1,7 @@
+carla_dataset:
+  _target_: asim.dataset.dataset_specific.carla.carla_data_processor.CarlaDataProcessor
+  _convert_: 'all'
+
+  splits: ["carla"]
+  log_path: "${oc.env:HOME}/carla_workspace/data"
+  output_path: ${asim_data_root}
