+++ asim/script/config/datasets/nuplan_dataset.yaml
@@ -1,0 +1,7 @@
+nuplan_dataset:
+  _target_: asim.dataset.dataset_specific.nuplan.nuplan_data_processor.NuplanDataProcessor
+  _convert_: 'all'
+
+  splits: ["nuplan_train", "nuplan_val", "nuplan_test"]
+  log_path: ${oc.env:NUPLAN_DATA_ROOT}/nuplan-v1.1/splits # NOTE: folder including [mini, trainval, test], sometimes not inside "splits" folder
+  output_path: ${asim_data_root}
