+++ asim/script/run_dataset_caching.py
@@ -1,9 +1,11 @@
 import logging
+from typing import List
 
 import hydra
 from nuplan.planning.script.builders.logging_builder import build_logger
 from omegaconf import DictConfig
 
+from asim.script.builders.data_processor_builder import RawDataProcessor, build_data_processor
 from asim.script.builders.worker_pool_builder import build_worker
 
 logger = logging.getLogger(__name__)
@@ -22,12 +24,15 @@ def main(cfg: DictConfig) -> None:
     build_logger(cfg)
 
     # Build worker
-    build_worker(cfg)
+    worker = build_worker(cfg)
 
     # Precompute and cache all features
     logger.info("Starting Dataset Caching...")
-    print(cfg.datasets)
-    # cache_data(cfg=cfg, worker=worker)
+    data_processors: List[RawDataProcessor] = build_data_processor(cfg.datasets)
+    for data_processor in data_processors:
+        logger.info(f"Processing dataset: {data_processor.__class__.__name__}")
+        data_processor.convert(worker=worker)
+        logger.info(f"Finished processing dataset: {data_processor.__class__.__name__}")
 
 
 if __name__ == "__main__":
