+++ scripts/dataset/run_log_caching.sh
@@ -2,5 +2,4 @@
 
 
 python $ASIM_DEVKIT_ROOT/asim/script/run_dataset_caching.py \
-+datasets='["carla", "nuplan_mini_train"]' \
 experiment_name="caching"
